requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pillow==9.5.0
werkzeug==2.3.7
numpy==1.24.3
requests==2.31.0
celery==5.3.4
redis==5.0.1
google-cloud-vision==3.7.2
pandas==2.2.2
fuzzywuzzy==0.18.0
chardet==5.2.0
google-generativeai==0.5.2
python-dotenv==1.0.1
pdf2image==1.16.3